chore: Remove deprecated non-paginated Find APIs
## Description

Removes various "Find" APIs which did not support pagination.
These had been marked deprecated nearly a year ago, and are not in use by Forge Cloud anymore.

Note that ssh-console still used some of the methods - therefore a deployment of a new ssh-console will be required when this Carbide release is rolled out.

## Type of Change

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [x] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

## Breaking Changes
- [x] This MR contains breaking changes

New ssh-console version needs to be deployed in combination with Carbide

## Testing

- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4848