feat(admin,api,rpc,scout): DPA + mlx device management w/ scout-stream
This pulls together all of the work that has been done on developing the `mlxconfig-*` crates for doing Mellanox device management, and introduces an administrative interface, via a new `ScoutStream` mechanism, that can be accessed via `forge-admin-cli` (and eventually the UI, but I'm not doing that in this MR).

The crates themselves are going to be used as part of the automated DPA provisioning workflows for Mellanox devices, starting with DPAs, which includes the work in !4730.

However, we need a way to be able to manage these devices administratively, outside of the automated provisioning workflows, without needing to log in to machines and understand how to use the tooling. This supports cases where the automated provisioning flows are having problems, causing problems, or need us to intervene and/or troubleshoot issues.

To make this an enjoyable (and complete) user experience, I realized using the CLI to manage devices wasn't going to really work with an agent polling mechanism as the plumbing, so I'm introducing a new bidirectional streaming endpoint that exists for `scout` called `ScoutStream`. This is a single endpoint between `scout` and `carbide-api` (with RBAC rules to only allow `scout`), which leverages a `oneof` to allow bidirectional message passing of different message types over the stream. Some high level actions include being able to see current `ScoutStream` connections (as well as basic `ping` and `disconnect`) admin calls, but, more importantly:
- Being able to get direct Mellanox device reports.
- Being able to lock/unlock Mellanox devices directly.
- Being able to apply `mlxconfig` profiles to Mellanox devices directly.
- Being able to change individual `mlxconfig` variables directly.

```
                                                        ┌────────┐
                                            ┌───────────│admin ui│
                                            ▼           └────────┘
┌────────┐      ┌───────────────┐    ┌─────────────┐
│ scout  │◀────▶│ scout-stream  │◀──▶│ carbide-api │
└────────┘      └───┬───────────┤    └─────────────┘
                    │  oneof A  │           ▲
                    ├───────────┤           │   ┌────────────────┐
                    │  oneof B  │           └───│forge-admin-cli │
                    ├───────────┤               └────────────────┘
                    │  oneof C  │
                    └───────────┘
```

These are all things which will be managed by the automated DPA flows, but this gives us the ability to manage them using `forge-admin-cli` (and again, the UI eventually if we want).

ALSO, I would like to eventually fold additional calls into the streaming model. All existing `scout` <-> `carbide-api` flows are untouched, but we could slowly move things to a single endpoint over which all traffic flows.

ALSO, we can add other things, e.g. we could add `forge-admin-cli scout-stream enumerate <machine-id>` if we wanted to get full hardware info directly from the machine.

Tested fully end to end in my `local-dev` environment.

Note: For Mellanox devices, you can manage them via `scout` if they are either SuperNICs, or DPUs in NIC mode. If it's a DPU in DPU mode, you'll be locked out entirely. In this case, we would also connect `forge-dpu-agent` to this system (or a similar stream, specifically for DPUs), allowing us to target those as well.

## Type of Change
<!-- Check one that best describes this MR -->
- [X] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)
<!-- If applicable, provide JIRA or/and NVBugs ticket. If none, the MR itself will be referenced in changelog -->
JIRA: [FORGE-5987](https://jirasw.nvidia.com/browse/FORGE-5987)

## Breaking Changes
- [ ] This MR contains breaking changes

<!-- If checked above, describe the breaking changes and migration steps -->

## Testing
<!-- How was this tested? Check all that apply -->
- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes
<!-- Any additional context, deployment notes, or reviewer guidance -->

See merge request nvmetal/carbide!4833