FORGE-7330: ignore error when a DPU returns an error getting ethernet interfaces; provide and use info in pxe boot request to distinguish a DPU
## Description

way back when, a PXE request for an ARM meant DPU, but when ARM hosts arrived, this changed because they have different boot images.  We then changed it to rely on site-explorer to determine the machine type, but if the DPU doesn't return ethernet interface information, it doesn't work.

allows a DPU to boot if there is no associated machine record.  It request that the DPU provide "platform" info in the pxe request.

## Type of Change

<!--Check one that best describes this MR-->

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality
- [x] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

FORGE-7330

## Breaking Changes

- [ ] This MR contains breaking changes

<!--If checked above, describe the breaking changes and migration steps-->

## Testing

<!--How was this tested? Check all that apply-->

- [x] Unit tests added/updated
- [ ] Integration tests added/updated
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

<!--Any additional context, deployment notes, or reviewer guidance-->

See merge request nvmetal/carbide!4815