Add a force option to unassign a SKU
## Description

adds a `--force` flag to the `sku unassign` command.  while not recommended its possible that a machine may be stuck in a state and unassigning the SKU would be helpful.

## Type of Change

<!--Check one that best describes this MR-->

- [ ] **Add** - New feature or capability
- [x] **Change** - Changes in existing functionality
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

<!--If applicable, provide JIRA or/and NVBugs ticket. If none, the MR itself will be referenced in changelog-->

JIRA: NVBugs:

## Breaking Changes

- [x] This MR contains breaking changes

An old CLI would not be able to unassign a sku with an updated API.  Since this is mostly used by SRE and they take the cli from the container, I don't feel this is a big deal.

## Testing

<!--How was this tested? Check all that apply-->

- [ ] Unit tests added/updated
- [ ] Integration tests added/updated
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

<!--Any additional context, deployment notes, or reviewer guidance-->

See merge request nvmetal/carbide!4907