feat: Add iteration_id and enqueued_objects tables for state controllers
## Description

This is the first MR in a set of MRs that aim to make the state handler framework "faster" and "more dynamic".
The general idea is that instead of scheduling the state handler only at a fixed frequency (once every 30s), it can run more often, but with at least the guaranteed fixed frequency.

In order to get there the following concepts are introduced:
- A periodic task inside the state controller schedules all objects for execution. They are stored in a new table `DB_QUEUED_OBJECTS_TABLE_NAME` (introduced in this MR).
- In addition to the periodic task, other workflows can immediately schedule the same object for execution again by writing to the same table.
- The state controller will cont continuously dequeue resources from the table and schedule their state handler execution.

Within this change, we only introduce two new database tables:
- `DB_ITERATION_ID_TABLE_NAME`: Uniquely identifies the current run of a state handler. This will later on be required to determine whether the handler had been executed in a certain run before, or whether its newly executed.
- `DB_QUEUED_OBJECTS_TABLE_NAME`: Is used to schedule objects for execution

Follow-up changes will guarantee that the handlers are more immediately scheduled.

## Type of Change

- [x] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

JIRA: First part of https://jirasw.nvidia.com/browse/FORGE-3039

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing

- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4921