Custom lint to warn when holding database transactions across unrelated async work
## Description

When perf testing carbide with 5,000 hosts (10,000 when including DPU's),
ingestion with a high degree of parallelism ends up causing a large number of
connection slots to be used up in the database connection pool, and there ends
up being a lot of "idle in transaction" processes in postgres, which is a sign
that a transaction is open but without progress being made.

We generally fix this by limiting the concurrency so that only so many hosts
are ingested at once, but this makes ingestion much slower than it needs to be.

On top of this, if there are systemic connectivity issues with BMC's, "idle in
transaction" connections can skyrocket and occupy all of the available
connection slots.

The pattern we want to avoid, is ever keeping a postgres transaction open while
we do other unrelated work, like redfish calls, infiniband, etc.

This adds a lint, `txn_held_across_await`, which will warn if you're ever doing
this. A detailed description is in the [README file][readme], but it is called
via `cargo carbide-lints`, and warns if a `.await` happens while a
`sqlx::Transaction` (or `db::Transaction`) is in scope, *and* if that `await`
isn't for a function that borrows the transaction (ie. database functions
themselves, or a wrapper.)

Finally, this adds `#[allow(txn_held_across_await)]` in 22 different locations 
where we are currently violating this lint... this is our TODO list on fixing
this issue.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)


## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [X] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

[readme]: https://gitlab-master.nvidia.com/nvmetal/carbide/-/blob/lint-txn-held-across-await/lints/carbide-lints/README.md

See merge request nvmetal/carbide!4937