Internal: Refactor recent dev-env test dependency
## Description

Recently the pre-merge dev-env test was given a dependency on the bfb boot artifact. This was achieved in the pipeline by adding a gitlab 'needs' dependency, but this doesn't follow the existing pattern. 

Refactoring to use the bfb boot artifact that is built during the test job, and removing the dependency on previous pipeline job.

Note: this is not the desired end state. Ideally we wouldn't re-build all the artifacts during the test, and instead pull/inherit the images that have just been built by previous jobs. But that's a non-trivial change.

## Type of Change

<!--Check one that best describes this MR-->

- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [x] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

<!--If applicable, provide JIRA or/and NVBugs ticket. If none, the MR itself will be referenced in changelog-->

JIRA: NVBugs:

## Breaking Changes

- [ ] This MR contains breaking changes

<!--If checked above, describe the breaking changes and migration steps-->

## Testing

<!--How was this tested? Check all that apply-->

- [ ] Unit tests added/updated
- [ ] Integration tests added/updated
- [x] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

<!--Any additional context, deployment notes, or reviewer guidance-->

See merge request nvmetal/carbide!4944