chore: Emit statehandler counters and histograms for each object immediately
## Description

This change is part of a series that will make it possible to run state-handlers for objects more often than with just a fixed frequency. In order to get the counters and histograms for all handling attempts emitted, we directly emit them from the state handler task.

The change removes the `times_in_state_s` and `handler_latencies_us` fields from the state handler log span because the values are no longer available after the factoring. However the values do not seem to have been used up to now.

## Type of Change

- [ ] **Add** - New feature or capability
- [x] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Related Tickets (Optional)

## Breaking Changes
- [x] This MR contains breaking changes

The change removes the `times_in_state_s` and `handler_latencies_us` fields from the state handler log span because the values are no longer available after the factoring. However the values do not seem to have been used up to now.

## Testing

- [x] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4954