Also disallow PgConnection in txn_held_across_await
## Description

nvmetal/carbide!4937 added a custom lint to warn if a transaction is being held
across an await point, if that await point is doing work unrelated to making
database queries.

But in most of the "middle" places that accept a transaction and then await
other functions with it, we're not taking a `PgTransaction`, but a
`PgConnection`, causing false negatives.

This means the vast majority of the time we're holding a PgConnection, we're
really holding a transaction, so it's best to just treat any PgConnection as
potentially holding a transaction. This requires allowing txn_held_across_await
in 34 more places, mostly in preingestion manager and the machine state
handler.

The general fix plan for this will be to pass a `PgPool` around directly, which
can be passed as an argument to sqlx functions as it implements the sqlx
Executor trait as well. This would mean that connections are taken lazily from
the pool as we make queries, which is what we want. But this will require a lot
of testing, so just allow the lint for now.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4958