Fix issue with transaction scoping in carbide-lints
## Description

I found some new false positives while doing a big refactor of the way we're
holding transactions, and it seems to be related to how the MIR search for
transaction locals produces spans that don't compare equal to the spans in the
coroutine witness info (even though they refer to the exact same span of code.)

While trying to fix it, I changed the approach to detecting "is this
transaction being passed to the awaited statement" to match the exact local
that's being held across the await, rather than matching on any transaction
local, and in the process I ended up using the span returned from
find_mir_local, which fixed the issue.

This approach turns out to be more robust (and actually fixed some false
negatives that went unnoticed in the test, as well as some duplicate warnings),
so here it is.

The result is that nested transactions are now errors, since technically the
outer transaction is being held while the inner transaction is passed... this
is likely going to stay using `#[allow]`, as it's a very non-trivial fix.

Several false positives are now fixed as well.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [X] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes
Builds on nvmetal/carbide!4958, look at the last commit to see what's in this MR.

See merge request nvmetal/carbide!4977