Fix AsyncFirmwareUploader to not need awaits
## Description

In nvmetal/carbide!4958, I have a fix for a the `txn_held_across_await` lint that
ensures it fires for functions holding a `PgConnection` instead of just a
`PgTransaction`. The MR tags `HostUpgradeState::waiting_for_upload` with
`#[allow(txn_held_across_await)]` to allow the behavior there, but it only needs
it because it was awaiting a tokio mutex. But we don't need tokio mutexes here,
we only need regular ones.

(tokio mutexes are for when you *need* to await something while holding a lock.
This is almost always a bad idea. If you don't need to await something while
holding a lock, you should use std::sync::Mutex, which will fail to compile if
you ever do await while holding it.)

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes

There's a lot of MR's waiting to merge, this will conflict with !4958 or !4977,
which both introduce this `#[allow]`. I'd prefer to let (one of) those merge
first, then fix the conflicts here.

See merge request nvmetal/carbide!4979