fix another false positive in txn_held_across_await
## Description
There's no issues with it in trunk yet, but in a branch where I'm
working on eliminating long-running transactions, I found a case where
this fires:

```
let mut txn = pool.begin().await?;

if some_condition {
    txn.rollback().await?;
    let mut txn = pool.begin().await?; // different local
    do_something_with(txn).await?; // false positive here
}

// txn is still "maybe" alive here
```

This is due to the liveness checking seeing if the variable is dead at
the await point, but it never actually found the await point, because
the span it was looking for wasn't comparing equal with `.eq()`. So it
went all the way to the end of the function, at which point txn was
"maybe alive" again, and thus the lint fired.

Fix this by using `Span::source_equals`, which is meant for the scenario
where Spans are produced by different layers in rustc and will fail
equality even if they point at the same source. Comments for
source_equals:

> Returns true if the spans are equal with regards to the source text.
>
> Use this instead of == when either span could be generated code, and
> you only care that they point to the same bytes of source text.


## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [X] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [ ] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4993