txn_held_across_await: Find lints inside async closures as well
## Description
This fixes the last false-negative I know about in the txn_held_across_await
lint: Using a transaction inside an `async [move] { ... }` closure.

For one, we need to explicitly crawl the HIR body of each function looking for
closures, and pass them to the lint as well.

But even with that, the part that took a long time to figure out: It turns out
for closure upvars, you need to check `tcx.closure_captures()` to find any
variables captured by the closure. But it's a dead-end trying to find
`LocalDecl` information from those, since there are none.  Instead, you have to
use the "Place" (rust's word for an lvalue) of each capture directly when
calculating liveness information.

So this bifrucates the lint a bit:

- For async fn's, we use the coroutine layout info, which gives us spans for
  each local being held across an await, and we check "is this await allowed" by
  finding LocalDecl's for each txn span, and checking if those LocalDecl's are
  alive.
- For async fn's, we instead get Place information, which means we check "is
  this await allowed" by finding the HirId of each txn directly, and by checking
  if the Place is alive.

Each of these cases are really hard to unify, since the two ways to get the info
we need are so different... so for now, just split things into multiple
functions. It's ugly, but it works for now. (At some point I want to put the
effort into seeing if I can get Place information directly for the async fn use
case, and unify this somewhat. For now, it doesn't seem to work, while this
approach does.)

In the process, I realized I was only checking two categories of items
(ItemKind::Fn and ImplItemKind::Fn) to scan for HIR bodies, and that there's
already a helper for this, Node::associated_body, which finds other types of HIR
bodies. This added a bunch of new lint warnings, and I had to add more allows to
each of them. (The lint was missing fn bodies defined in `impl X for Y` before.)


## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [X] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

I should have some MR's soon for the actual work I've been doing actually fixing
the issues found by this lint... but I just keep finding issues like these with
the lint as I go, so fixing the lint comes first.

See merge request nvmetal/carbide!5017