Fixing transactions held across await points, part 1/n
## Description

This fixes one of the more minor cases of transactions being held across
await points (see nvmetal/carbide!4937 for background), where we're
holding an advisory lock in postgres to make work mutually exclusive.

There's probably half a dozen of these use cases across the codebase...
they don't add up to a lot of open transactions, but they can cause the
transaction to be held a long time in the event that the work is slow,
so it's worth fixing.

This approach makes a table, `work_locks`, to indicate that a given
piece of work is in progress, and uses RAII-style drop traits to release
the lock when the work is done. In the event the process crashes or is
forcefully killed, the locks will expire after a timeframe so that other
workers can pick up the work.

I'm using preingestion manager as a random example to use WorkLockManager,
after this merges I have more branches that port the other cases over.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [X] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!4991