Use PgTransaction instead of PgConnection when we do locks
## Description
While testing my txn_held_across_await fixes, there were a few cases where I
wanted to simply acquire a connection instead of starting a transaction (to
avoid needing to remember to commit if we know we're just going to write to one
thing, etc), and came across cases where we issue LOCK TABLE statements, but
only requiring that the caller pass a PgConnection.

Postgres throws an error if you try to LOCK TABLE when you're outside of a
transaction, so we should make these functions take an actual PgTransaction
instead of a PgConnection (a PgTransaction can deref_mut into a PgConnection but
not vice-versa) so that the type system helps us enforce that invariant.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [ ] Integration tests added/updated  
- [ ] Manual testing performed
- [X] No testing required (docs, internal refactor, etc.)

## Additional Notes

See merge request nvmetal/carbide!5021