Use work manager everywhere
## Description

As part of fixing all instances of txn_held_across_await, we can't use
transactions to lock work and then do a bunch of slow work involving redfish
calls, etc.

So for things like state controller IO, IbFabricMonitor, MachineUpdateManager,
etc, take all of the cases where we used to do an advisory lock on a table, and
change it to use WorkLockManager, which was introduced for PreingestionManager
in nvmetal/carbide!4991.

## Type of Change
- [ ] **Add** - New feature or capability
- [ ] **Change** - Changes in existing functionality  
- [ ] **Fix** - Bug fixes
- [ ] **Remove** - Removed features or deprecated functionality
- [X] **Internal** - Internal changes (refactoring, tests, docs, etc.)

## Breaking Changes
- [ ] This MR contains breaking changes

## Testing
- [ ] Unit tests added/updated
- [X] Integration tests added/updated  
- [X] Manual testing performed
- [ ] No testing required (docs, internal refactor, etc.)

## Additional Notes

The next step after this is to stop passing the txn object to the individual
state controller IO impl's and pass a PgPool instead, letting them manage their
own transactions (avoiding interleaving them with slower calls.) I have this in
a branch but am trying to avoid putting up too huge of a change, to ensure
reviews are manageable.

See merge request nvmetal/carbide!5022